Add coverage for subprocess specs
Since PR https://github.com/rubocop-hq/rubocop/pull/6255 the coverage results
from the different suites (one per tested Ruby version) are merged together, and
the result is sent to Code Climate.

That means we no longer need to ignore coverage results from the specs that are
run in a separate process. We just need to give these coverage results their
own name (e.g. a namespace) and set a new PID (I think to stop SimpleCov's
`at_exit` hook when the fork ends) and these subprocess results will be merged
with the coverage results of the rest of the suite.