Clarify dangers of Rails/LexicallyScopedActionFilter
This cop encourages users to define action methods in the same class as
any before_, after_ or around_actions that refer to the action.

However, the examples given don't account for the fact that behaviour
might be defined in the superclass. Indeed, this sort of decoration is
one of the main use cases for action decorators.

If this is the case, then the empty action definitions shown in the docs
will override the desired behaviour, and lead users to introduce bugs.
This change emphasises this in the docs, and gives an example to
encourage users to think about the inheritance chain.