[Fixes #6939] Remove warn for rubocop-performance
Fixes #6939.
Fixes #6917.

This warning should not be displayed if users don't use Performance cops
as follows:

- https://github.com/rubocop-hq/rubocop/pull/6845#issuecomment-481810841
- https://github.com/rubocop-hq/rubocop/pull/6919#issuecomment-483005401

In the future, it will be better to replace it with a mechanism that
displays a warning only when the user need Performance cops.

This PR removes a warning for rubocop-performance when runinng `rubocop`
command.

It just leave a warning just to `gem install rubocop`.
https://github.com/rubocop-hq/rubocop/blob/v0.67.2/rubocop.gemspec#L47