Fix a false positive for `Layout/EmptyLinesAroundAccessModifier`
This PR fixes a false positive for `Layout/EmptyLinesAroundAccessModifier`
when at the end of specifying `self` is missing blank line.

It is a regression with the following commits.
https://github.com/rubocop-hq/rubocop/commit/4d709506a26bf3ed00aca7a2898db426d3bdb79a

This PR fixes an infinite loop similar to #6576.

This issue occurred only on master branch, it will not be described
in the changelog.