[Fix #7063] Fix unsafe autocorrect for 'Style/TernaryParentheses'
This takes the precedence of operators used into account when deciding
to autocorrect. The issue is that if an operator of precedence below
that of the ternary operator (e.g. "or") is used, then the existing
autocorrect can change the semantics by removing parentheses. With
this change, the autocorrect will no longer attempt to make this
correction.