Fix Layout/EmptyLinesAroundBlockBody for multi-line method calls
This cop currently looks for blank lines following the first line of the
block node. When the block is passed to a multi-line method call, the
send node spans multiple lines, and the wrong line is inspected.

This is a similar problem to 0c569e32fa51cf4650def5d214fc00a5e8077381,
where a class definition can span multiple lines.