Change logic for when to use EOL disable comment
The decision whether to use a single EOL comment to disable a cop, or a
disable/enable pair on separate lines, should not be based on whether the
code range of the offense is single line or multi-line. It should be based
on whether the comment fits on the given line.

The single line comment WILL disable the offense even if it's reported on a
multi-line range.

We also need to make sure we don't add more than one EOL comment per line
(fix later).