Allow `#**` Doxygen comment style on `Layout/LeadingCommentSpace`
This PR remove the offense for comments that starts with `**` as this style is
a valid syntax to write block comments in Ruby.

Motivation: the `apidocjs` package only specify `=begin/=end` block to define
the documentation in ruby. However, [this comment][1] explain that the `Doxygen`
style `#**` is also compatible to write the documentation so the `Layout/LeadingCommentSpace`
should not flag any offense for it.

My [first attempt][2] was to not flag `Style/BlockComments` if the `@api` annotation
was found inside the `=begin/=end` block but this solution is more elegant.

Example:

```ruby
=begin
@api {get} /endpoint description of my endpoint
...
=end
```

```ruby
- #**
- # @api {get} /endpoint description of my endpoint
- # ...
- #*
```

[1]: https://github.com/apidoc/apidoc/issues/620#issuecomment-494206190
[2]: https://github.com/rubocop-hq/rubocop/pull/6522