[Fix #3534] Report long modifier lines in IfUnlessModifier
Make Style/IfUnlessModifier work in two directions. Apart from the already
existing reporting of if/unless expressions on normal form that could be
changed to modifier form, we add reporting and correcting of modifier form
if/unless expressions that make the line too long and should be written on
normal form.