[Fix #4802] Detect all UnneededCopEnableDirective offenses
After all other cops are finished we invoke the
Lint/UnneededCopDisableDirective cop, since it depends on the results of the
other cops. If it's an --auto-correct run, correcting
UnneededCopDisableDirective offenses can introduce new
UnneededCopEnableDirective offenses. For this reason we must do one more
inspection after UnneededCopDisableDirective has corrected a file.