[Fix #7355] Print warning for cop names given without department
Ever since departments (namespaces) were introduced, or almost since then, it's
been possible to give cop names on the command line, in configuration and in
rubocop:disable comments without a preceding department name. RuboCop deduces
what the department should be and errors out only if it doesn't find exactly
one matching department.

The optional department functionality was added as a way to stay backwards
compatible with configuration that was created before departments existed. But
the grace period is coming to an end. We start now by issuing warnings whenever
cop names without departments are encountered. Later on, in a future version,
we will remove the department deduction functionality and require that
departments always be given for cop names.