Fix heredoc edge cases in Layout/EmptyLineAfterGuardClause
If you call a method on the heredoc string the cop didn't couldn't
recoginze that a heredoc is passed. The same happened if the heredoc was
given to a method, the result of which is passed as a value in another
call.

Here are the two examples:

```ruby
def foo
  raise ArgumentError, <<~END.squish.it.good unless guard
  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Add empty line after guard clause.
    A multiline message
    that will be squished.
  END

  return_value
end
```

```ruby
def foo
  raise ArgumentError, call(<<~END.squish) unless guard
  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Add empty line after guard clause.
    A multiline message
    that will be squished.
  END

  return_value
end
```