[Fix #7426] Add `always_true` to Style/FrozenStringLiteralComment
This adds another style to Style/FrozenStringLiteralComment that
enforces that the comment is always enabled, that is, it requires that
string literals are always frozen. The default style, `always`,
enforces that the comment exists but does not enforce that the comment
is enabled.

This also differentiates the offense message for the `always` and
`always_true` styles.

It changes the offense message when using the `always` style to be
less specific, since the old version of the message implied that the
only acceptable value of the frozen_string_literal comment was
`true`. In fact, in the `always` style, either `true` or `false` is
acceptable. The new `always_true` style uses that preexisting message,
since when using that style it is required that the value of the
frozen_string_literal comment is `true`.