Added new cop - Naming/AsciiConstants
This PR adds a new cop to add offenses to constants with non-ascii characters
in their declaraction. Currently this exists for identifiers in the
"Naming/AsciiIdentifiers" cop, however this only applies to tIDENTIFIER
and not tCONSTANT.
With this change: class / module / constant declaractions must contain only
ASCII characters to be successful.