Change terminology to `Allowlist` and `Denylist`
### Summary

Follow up #6464, #6466, and #6467.

This PR changes a terminology from `Whitelist` and `Blacklist` to
`Allowlist` and `Denylist`.

This change is an obvious breaking change to some cop options.
So I'd like to introduce it before RuboCop 1.0 if this change is acceptable.

### Other Information

This change has also been made in Rails repository and other repos.
https://github.com/rails/rails/pull/33681