[Fix #7506] Move LineLength code to mixin and use in IfUnlessModifier
Extract some code that can detect lines that are too long with all the special
cases that are controlled through configuration parameters of the
Metrics/LineLength cop.

The code that is moved to the LineLengthHelp mixin can then be used in
IfUnlessModifier so that it only reports lines on modifier form that are too
long when the LineLength cop would report them.