Move `LineLength` cop from `Metrics` to `Layout` department
### Summary

The `Layout` department was created after `Metrics` was created.

- Create namespace Metrics and move some cops there
  https://github.com/rubocop-hq/rubocop/commit/e2153e8cdaea8a79ab51977b0071c7b15cf95078
- Layout department
  https://github.com/rubocop-hq/rubocop/pull/4278

As far as I use `LineLength` cop, I think this cop belongs to the layout,
not the logic metrics. Rather than changing the logic when `LineLength`
cop becomes a offense, a solution is often a line wrap.

The horizontal length targeted by `LineLength` cop is resolved by
often a line wrapping, while the vertical length targeted by
`ClassLength` cop, `MethodLength` cop, and `BlockLength` cop is
resolved by a logic change.
So `ClassLength`, `MethodLength`, and `BlockLength` cop remain in
the `Metrics` department.

### Other Information

I'm wondering if the default maximum length of 80 is reasonable.
This is because there are cases where clear variable name (and etc)
have to be omitted due to restrictions on `LineLength`.
However, this PR does not handle it because first of all this PR aims
to move department.