Exposing correctable status on offense and in formatters
We're working on a project to run Rubocop on a list of Chef Cookbooks
and report information back to the user. One of the fields we need to
expose in addition to 'corrected' is 'correctable'.

It seemed to me, looking at the logic, that the `:unsupported` status
is the only one which indicates it _cannot_ be autocorrected.

Signed-off-by: tyler-ball <tball@chef.io>