Add new Lint/NonDeterministicRequireOrder cop
Dir[...] and Dir.glob(...) make no guarantees about the order files
are returned in:

> Case sensitivity depends on your system [...], as does the order
> in which the results are returned.

This becomes a problem when they are used for applications in which
the order can matter, such as requiring files. At worst, this can
lead to bugs that only happen intermittently in production and can't
be reproduced in development.

This cop suggests adding a .sort when requiring the files:

> # bad
> Dir["./lib/middleware/*.rb"].each do |file|
>   require file
> end

> # good
> Dir["./lib/middleware/*.rb"].sort.each do |file|
>   require file
> end