In default.yml, Exclude tmp
In certain rails projects, `tmp` can contain hundreds of thousands of files.

Excluding `tmp` in my project seems to decrease `time bin/rubocop` from 90s to 20s, though that's just an anecdote and not a scientific benchmark.

Does anyone put code in `tmp` that they actually want to lint? That'd be rare, right?