[Fix #7586] Fix infinite loop in FrozenStringLiteralComment
It was detecting frozen_string_literal in the first three lines of the
file, while there's no guideline to separate magic comments with
newlines. If shebang and encoding were separated by a newline, that was
three lines already, and frozen_string_literal was added on line 4, and
not detected as such after auto-correction, triggering an infinite loop.