[Fix #7616] Fix an incorrect autocorrect for `Style/MultilineWhenThen`
Fixes #7616.

This PR fixes an incorrect autocorrect for `Style/MultilineWhenThen`
for when statement with then is an array or a hash.

The following is a reproduction procedure.

```console
% cat example.rb
case condition
when foo then {
    key: 'value'
  }
end

% ruby -c example.rb
Syntax OK
```

It is changed to the code with syntax error as follows.

```console
% bundle exec rubocop -a --only Style/MultilineWhenThen
(snip)

% cat example.rb
case condition
when foo {
    key: 'value'
  }
end

% ruby -c example.rb
example.rb:3: syntax error, unexpected ':', expecting '}'
    key: 'value'
example.rb:4: syntax error, unexpected '}', expecting end-of-input
```

The same is true when replacing a hash with an array.
The PR will accept these cases without offense.