[Fix #7620] Fix a false positive for `Migration/DepartmentName`
Fixes #7620.

This PR fixes the following false positive for `Migration/DepartmentName`
when a disable comment contains a plain comment.

```console
% cat example.rb
# rubocop:disable Style/SafeNavigation # support Ruby < 2.3.0

% rubocop --only Migration/DepartmentName
Inspecting 1 file
C

Offenses:

example.rb:1:50: C: Migration/DepartmentName: Department name is missing.
# rubocop:disable Style/SafeNavigation # support Ruby < 2.3.0
                                                 ^^^^

1 file inspected, 1 offense detected
```

The following is an example disable comment.

```ruby
# rubocop:disable Style/FrozenStringLiteralComment, Style/StringLiterals
```

The token used after `# rubocop: disable` are `A-z`, `/`, or `,`.
Also `A-z` includes `all`.

```ruby
# rubocop:disable all
```

This PR will change the processing of `Migration/DepartmentName` cop to
finish when an invalid token is used.