Fix an edge case in `omit_parentheses` style of `Style/MethodCallWithArgsParentheses`
Before this change:

```ruby
foo(a) || bar(b) do
             ^^^ Omit parentheses for method calls with arguments.
  pass
end
```

This resulted in a `SyntaxError`, if corrected. The check for a call
with logical operator did not take into account that the parent of a send node with a block is the block node itself.

After this change, the code below is lawful:

```ruby
foo(a) || bar(b) do
  pass
end
```