Enable `Migration/DepartmentName` cop by default
This PR enables `Migration/DepartmentName` cop by default.

The cop has been added at #7385.

The cop will promote the migration to adding department name to cop name.
Omitting a department name will not be supported in the future,
so enabling this cop by default will facilitate the migration.

The PR was inspired by the following issue:

> RuboCop no longer supports ambiguous cop names (those without a department).

https://github.com/rubocop-hq/rubocop/issues/7717#issuecomment-586571995