Change `AllowComments` of `Lint/SuppressedException` to true by default
## Summary

Follow https://github.com/rubocop-hq/rubocop/issues/7052#issuecomment-494360787.

This PR changes `AllowComments` option of `Lint/SuppressedException` to
true by default.

As is used in RuboCop repo itself, it is a common practice to use
source code comments to explain when exception handling is not performed.

## Related Information

I think that `Suppressing Exceptions` rule can be refined to a suitable example.
https://github.com/rubocop-hq/ruby-style-guide#dont-hide-exceptions