[Fix #7831] Fix a false positive for `Style/HashEachMethods`
Fixes #7831.

This PR fix a false positive for `Style/HashEachMethods` when there is no
receiver for `keys` and `values`.

The following is an examples.

```ruby
values.each { |k, v| do_something(k, v) }
keys.each { |k, v| do_something(k, v) }
```

False negatives occur in cases such as object that inherit Hash,
but that is perhaps a corner case.
I think it is more worthwhile to resolve the false positives with
common naming of `values`.