Add support for tabs in indentation (#7867)
This commit adds SupportedStyles to Layout/Tab - spaces and tabs. It also renames the cop to
to Layout/IndentationStyle, as the old name no longer reflects its purpose.

One notable difference is that before the `Layout/Tab` cop was warning on tabs anywhere in the source, but now we're focusing only on tabs used for indentation purposes.