[Fix #7790] Set options in ConfigLoader earlier
Transferring options to the ConfigLoader must be done before we
validate the --parallel flag against configuration. Since we cache
the configuration when reading it to perform validation, the
cached configuration will not take all the given options into
account, which leads to different behavior when running the actual
inspection if the --parallel flag is given.