[Fix #7881] Call act_on_options earlier in CLI#run()
This fixes another problem that's similar to the one reported in #7790. The
problem is still that configuration is read and cached when we validate options
vs configuration. By acting on all options first, we make sure that the first
time we read configuration, it's done according to the given options.
Specifically, it's the --force-default-config option that's now being
respected.