Fix a false positive for `Lint/LiteralAsCondition`
This PR fixes a false positive for `Lint/LiteralAsCondition`
when using `true` literal in `while` and similar cases.

`while true` is a known idiom.

Normally `loop` is used, OTOH `while true` may be selected for
speeding up. Users use these methods depending on use case.
That was introduced as part of RubyKaigi 2019.
https://rubykaigi.org/2019/presentations/ktou.html