[Fix #7885] Consider Layout/IndentationStyle:EnforcedStyle
When calculating how wide the indentation is, we must look at the
`EnforcedStyle` of the `Layout/IndentationStyle` cop. If it's `spaces`, then
the indentation multiplier is 1. If it's `tabs` it's a bit more complicated,
but that part was already handled correctly.

The specs needed to be updated with correct configuration for all cases with
`tabs` indentation.

Whether or not we should consider the `Enabled` parameter is debatable. I've
chosen to not care about `Enabled`, and to use the other parameter settings in
any case.