Add new Layout/PrivateConstant cop
This cop checks that there are no constants defined in the private or
protected blocks of a class. This does not actually make the constant
private and it will still be publicly accessible. The cop will ignore
constants that are explicitly marked with `private_constant`.

Addresses a closed issue raised in rubocop-hq/ruby-style-guide#536.