Add `EnforcedStyle` for `Style/DoubleNegation` cop
This PR adds `EnforcedStyle` option for `Style/DoubleNegation` cop,
and `EnforcedStyle: allowed_in_boolean_context` by default based on #7833 survey.

```ruby
# @example EnforcedStyle: allowed_in_boolean_context (default)
#   # good
#   def foo?
#     !!return_value
#   end
#
# @example EnforcedStyle: always
#   # bad
#   def foo?
#     !!return_value
#   end
```

Ruby will return the last evaluated value.

This PR only supports simple contexts, such as the end of line value of
method, and when using `return`.
A complex contexts should also be supported, such as when branching with
`if` and others in the future.