Set 100 as the internal Layout/LineLength Max
The decision to change the default Max value for Layout/LineLength was
based on a survey. People seem to prefer 120 as the limit for line length,
and that's fine. Let the people have the line length they prefer. But it's
not necessarily a good thing to immediately adopt the same configuration
setting internally in RuboCop. There's the "eat your own dog food"
argument, but one might also argue that if we wanted 120 character lines
in RuboCop we would have changed the internal setting earlier.

To me, the jump from 80 to 120 is too big. A bit of a shock actually. It
pains me to not be able to work with a vertically split screen and read
the source code without line wrapping occurring. I suggest 100 characters
as a middle ground. Following the RuboCop user community in the move to
longer lines, but not going all the way.