Add support for multi-line strings with backslash in InterpolationCheck
It is unclear why we should skip `str` nodes with a `dstr` parent in this
cop. Because that means we allow interpolation in single quoted strings as
long as they are split into several lines with backslash delimiters.

No existing test cases started to fail when this condition was removed.