Lazy enlisting
There's a problem in specs ...
I wouldn't be surprised if you stopped reading after this point, since it's usually not a worthwhile justification.
But please take a moment and bear with me.

If we define a cop for testing purposes like that:
```ruby
module Dept
  class MyCop < RuboCop::Cop::Cop
    # ...
  end
end
```
you'll end up with a constant littering the root namespace.
https://rspec.rubystyle.guide/#declare-constants
https://fili.pp.ru/leaky-constants.html

It's not a big deal if you pick unique class and namespace each time or are brave enough to deal with potential side effects.

The following is arguably better:
```ruby
klass =
  Class.new(RuboCop::Cop::Cop) do
    def external_dependency_checksum
      'something other than nil'
    end
  end
stub_const('Dept::MyCop', klass)
```

But there's a problem with this approach:

```
  1) RuboCop::Cop::Team#external_dependency_checksum when cop with different checksum joins has a different checksum for the whole team
     Failure/Error: new(*class_name.split('::').last(2))

     NoMethodError:
       undefined method `split' for nil:NilClass
     # ./lib/rubocop/cop/badge.rb:26:in `for'
     # ./lib/rubocop/cop/cop.rb:66:in `badge'
     # ./lib/rubocop/cop/registry.rb:35:in `enlist'
     # ./lib/rubocop/cop/cop.rb:62:in `inherited'
     # ./spec/rubocop/cop/team_spec.rb:372:in `initialize'
     # ./spec/rubocop/cop/team_spec.rb:372:in `new'
     # ./spec/rubocop/cop/team_spec.rb:372:in `block (4 levels) in <top (required)>'
```

Backtrace explains it all, but in a nutshell there's a pesky `def self.included` in `Cop` that enlists the newly defined cop for a department. And to get the department, it needs a name. Which is `nil` for `Class.new`.

Ha! Easy-peasy you think: define a class method `name` on that cop.
Not so easy it turns out:

```ruby
class A
  def self.inherited(base)
    fail 'Ooopsie!' unless base.name
  end
end

class B < A # ok
end

Class.new do # boom!
  def self.name
    'Multipassport!'
  end
end
```

And guess what? `self.inherited` is called **before** the block is evaluated.
`Class.new` to blame you would say!
But it's not. Here's another example for you.

```ruby
class A
  def self.inherited(base)
    fail 'Traitor!' unless base.surname
  end
end

class B < A
  def self.surname
    'Schtirlitz'
  end
end
```

```
NoMethodError: undefined method `surname' for B:Class
```

Srsly? WTH, Ruby? They have to quack, but some of them are real and have to have names and other things!

Ok, if it's not `Class.new` to blame, it's a premature `inherited` hook. What do we do?

Lazy enlist to the rescue!