Always loop when testing autocorrection
For most cops and specs, this won't change anything. The offense is
corrected in one go, and the second iteration doesn't change anything.

For a few specs this will change something. For a few specs, they have
to explicitly pass `loop: false` to test the result of the first round
of autocorrect, and in some specs we need to pass it to avoid exceptions
such as `InfiniteCorrectionLoop`. And that's the real value of using
looped autocorrection in the specs - we will find more errors, earlier.