[Fix #8083] Fix an error for `Lint/MixedRegexpCaptureTypes`
Fixes #8083.

This PR fixes an error for `Lint/MixedRegexpCaptureTypes` cop
when using a regular expression that cannot be processed by regexp_parser gem.
https://github.com/ammar/regexp_parser

Probably false negatives will be made, however false negatives will be
reduced when a processing pattern of regexp_parser increases.

So this PR relies on regexp_parser for error resolution.