Only look up to the project's root for exclusions
If there's a specific project configuration, we don't want to pick up
the exclusions in the user configuration. This doesn't happen already in
the case of XDG setups, or if the HOME directory is not a parent of the
project root directory. The specs currently configure such a setup, even
if not very common in real life, maybe to workaround this issue.

This PR fixes the problem by defining the root of project as the highest
folder up in the directory hierarchy that contains a `Gemfile` or
`gems.rb` file. Even if this definition matches most ruby projects, we
could add additional fallbacks (like a `.gemspec` file?) to include more
edge cases.