Add expect_offense mark for offense on empty line
Part of #8003

Use `^{}` to denote an offense at an empty line in `expect_offense`.

Example:

```ruby
expect_offense(<<~RUBY)

  ^{} Missing frozen string literal comment.
  puts 1
RUBY
```

Enable pending frozen string literal specs

Correct which line the offense is added to (always the first line).

As the first line is non-empty in these cases, use a normal caret `^` to
mark the offense.