Inline `ParserDiagnostic` module
Follow https://github.com/rubocop-hq/rubocop/pull/8395#discussion_r460202551.

This PR does a refactoring to inline `ParserDiagnostic` module.
It seems that it is simpler and easier to maintain if it is implemented
by each cop rather than mix-in the module.

I don't think `ParserDiagnostic` module is used outside of RuboCop core,
at least not in the RuboCop HQ products.
On the other hand, I added it to the changelog just in case for 3rd party gems.