Refactor if_unless_modifier_spec
Related to rubocop-hq/rubocop-rspec#862

This adds a `spaces` helper method and simplifies the `other_cops` `let`
blocks. There's still more that could be done to make these tests a
little easier to maintain, but thought this was a decent first step.
It's also probably easier to discuss in discrete steps.