Handle to_h with block in hash transformation cops
Since Ruby 2.6, `Hash#to_h` accepts a block which is called with each
pair and must return a key and value to include in the new hash.

If either the key or the value are always returned unchanged, using
`Hash#transform_keys` or `Hash#transform_values` is faster and clearer.

This is similar to the existing logic that detects `map { ... }.to_h`.