Revert "Mark Cops and Formatters as private"
This reverts commit 90d5333f4c150bfe33d07805a12879b7fa4d6201.

I was late to participate in the discussion below.
https://github.com/rubocop-hq/rubocop/pull/8501

Cop names are specified by users to .rubocop.yml, `--only`
option, and etc. IMHO, the names and behaviors cannot be
changed without impacting users.

Users should refrain from depend on it as an implementation API,
but I don't think it's hidden from users.