[Fix #8799] Fix SpaceInsideBlockBraces cop regression.
In 6c05f69f4c1919b40fe6e331bab97833cf87bd0e this cop
was changed in a way which prevents it from running multiple
times on a single line.

It appears that this is due to the misunderstanding about
what the 'unrecognized_style_detected' and 'opposite_style_detected'
methods are designed to do.

I do not believe that they should be used as inquiry methods,
but should be used as side-effecting methods.

e.g. `opposite_style_detected` ALWAYS returns a truthy value.