[Fix #8871] Fix a false positive for `Style/RedundantBegin`
Fixes #8871.

This PR fixes a false positive for `Style/RedundantBegin`
when using `begin` for method argument or part of conditions.

`begin` keyword may be redundant when using only one expression
in `begin` of each issue case.
However, since it is the unintended case for #8822, I think it
can be implemented as an enhancement different from this bug fix.