Use `public_send` instead of `send` when public sendable
In Ruby, there seems to be a trend to separate `send` into
`public_send` and `__send__`.
https://github.com/ruby/ruby/commit/3198e7ab

This PR changes to `public_send` to suppress unnecessary scope to
(private, protected) method calls. On the other hand, this PR will not
make the change to `__send__` because the change to `__send__` is premature
https://bugs.ruby-lang.org/issues/17100