[Fix #8966] Enforce no spaces in empty parens.
Even with `style` set to `spaces`.

Regardless of if we like spaces in our parens or if we don't, empty parens should not have spaces.

This commit fixes `SpaceInsideParens` Cop when the `style` is `spaces` so that it doesn't require a space inside the parens. Moreover, it registers an offence when there _is_ a space inside empty parens.

Additionally, a test was added for the default `style` to ensure there are no spaces inside empty parens.

And finally, the docs were updated to show that spaces are not permitted with _either_ `style`.

Note, had to extra out `process_with_space_style` to avoid method length Cop offence.

Fixes https://github.com/rubocop-hq/rubocop/issues/8966.