Fix RSpec/StubbedMock offenses
I know that the RSpec/StubbedMock cop can be controversial, and I didn't
want to enable it and *always* enforce its recommended style.

But in all three cases this cop finds in our spec suite, the mock
(defined by `expect`) is entirely unnecessary - using a stub (with
`allow`) is enough. If any of the stubs are changed, other expectations
in each example will fail.