[Fix #9023] Mark unsafe for `Style/CollectionCompact`.
Fixes #9023.

This PR marks unsafe for `Style/CollectionCompact` by default because
false positives may occur in the nil check of block arguments to the
receiver object.
For example, `[[1, 2], [3, nil]].reject { |first, second| second.nil? }`
and `[[1, 2], [3, nil]].compact` are not compatible. This will work fine
when the receiver is a hash object.