Allow `enums` by default for `Lint/ConstantDefinitionInBlock`
Follow https://github.com/rubocop-hq/rubocop/issues/9025#issuecomment-726203458.

This PR allow `enums` method by default for `Lint/ConstantDefinitionInBlock`.

#9027 was worried about false negatives for users who have defined constants
in user custom `enums` method's block. But it would be worthwhile to be
able to resolve this realistic issue for Sorbet by default rather than not
knowing if it would happen or not. And there is not much imagination that
the false negatives will occur.