Deprecate `EnforceSuperclass` module`
RuboCop core depended on this module when it supported Rails
department. Rails department has been extracted to RuboCop Rails gem.
I forgot to remove it when extracting Rails department from the core.

rubocop-hq/rubocop-rails#390 is ready to be removed from RuboCop core.
First of all, as soft deprecation, the document only indicates
that it is a deprecation warning.
After releasing RuboCop Rails several times, I will display a
deprecation warning and will remove it in RuboCop 2.0.