[Fix #8857] Modify --auto-gen-config regarding Exclude and inheritance
When we generate Exclude properties in .rubocop_todo.yml, we already check if
the same Exclude appears in existing configuration. The file list is then
copied into the generated file. This doesn't work as intended for Exclude
properties in custom configuration. They override (by default) the same
properties inherited from .rubocop_todo.yml, resulting in a configuration that
will not hide all the offenses when rubocop is next run. The expected behavior
is that --auto-gen-config should hide all offenses.

This change makes rubocop --auto-gen-config warn about this situation when it
occurs. If inherit_mode is set to merge for Exclude properties, the behavior is
pretty much as expected today, but we tweak the behavior here too, and skip the
copying of file names that will be excluded anyway due to the merging.