[Fix #7726] Handle method call chains inside square brackets
Searching for the root of a method call chain should only consider method
calls with a dot operator. We should stop if it's another kind of send
node.

This is an improvement on a recent fix for the same cop. That commit
only fixed the problem for the `[]=` operator. This solution is more
comprehensive.