Improve the gemspec to load only the necessary files without the git utility
This makes the gemspec independent of the git utility by replacing the
files attribute value with Ruby Dir for listing all the required files.
In addition, we exclude bin folder as these files are strictly development
concerns and can cause issues when distributed. The same goes for the
assets which are unnecessary for the distributed package. Both the README.txt
and LICENSE.txt are no longer mentioned as well as they are automatically
included in the generated file list via the extra_rdoc_files attribute.
This change reduces the overall package size from 440K down to 432K.

The version file is now loaded with the require_relative method to
increase performance and skip unnecessary lookup.