Support regexp_parser 1.8 and 2.x series
Follow https://github.com/rubocop-hq/rubocop/pull/9154,
https://github.com/rubocop-hq/rubocop/issues/9155,
and https://github.com/rubocop-hq/rubocop/pull/9102.

This is a step towards the widespread use of regexp_parser 2.0.

RuboCop core accepts regexp_parser 1.8, but several code is already
incompatible with regexp_parser 1.8. It can cause issue because these
combination of versions.

Therefore, this PR ports code for regexp_parser 1.8 that will never be
maintained from https://github.com/rubocop-hq/rubocop/pull/9102.

Implementation of this patch, code is intentionally duplicated because
it is evaluated only when the class is defined. Also, since obsoleted
code for regexp_parser 1.8 is assumed to never be maintained, the target
to be removed is clear.

To be honest, it's ugly as implementation, but I think it has the least
impact for RuboCop 1.x series users.