[Fix #9234] Fix an error for `Style/KeywordParametersOrder`
Fixes #9234.

This PR fixes the error for `Style/KeywordParametersOrder` and
makes it aware of block keyword parameters.
Keyword parameters can be used with block arguments as well as method
arguments, but `Style/KeywordParametersOrder` cop were unaware of it.

This PR adds block arguments to the targets for consistency.
If the cop need an option to the above customize, I will open a separate PR.